feat: add RegistrationProfileWithDomainBlock
